+++ legacy_packages/wallets/test/smart-wallet-integration.test.ts
@@ -116,8 +116,9 @@ describeIf(!!SECRET_KEY)(
       expect(balance.toNumber()).toEqual(7);
     });
 
-    it("can sign and verify 1271 old factory", {
-      timeout: 240_000,
+    // FIXME flaky test
+    it.todo("can sign and verify 1271 old factory", {
+      timeout: 300_000,
     }, async () => {
       const message = "0x1234";
       const sig = await smartWallet.signMessage(message);
@@ -139,8 +140,9 @@ describeIf(!!SECRET_KEY)(
       expect(isValidV2).toEqual(true);
     });
 
-    it("can sign and verify 1271 new factory", {
-      timeout: 240_000,
+    // FIXME flaky test
+    it.todo("can sign and verify 1271 new factory", {
+      timeout: 300_000,
     }, async () => {
       smartWallet = new SmartWallet({
         chain,