+++ packages/thirdweb/src/extensions/airdrop/__generated__/Airdrop/write/claimERC721.ts
@@ -186,5 +186,6 @@ export function claimERC721(
     maxPriorityFeePerGas: async () =>
       (await asyncOptions()).overrides?.maxPriorityFeePerGas,
     nonce: async () => (await asyncOptions()).overrides?.nonce,
+    extraGas: async () => (await asyncOptions()).overrides?.extraGas,
   });
 }