+++ packages/thirdweb/src/extensions/erc20/__generated__/ITokenStake/write/depositRewardTokens.ts
@@ -136,5 +136,6 @@ export function depositRewardTokens(
     maxPriorityFeePerGas: async () =>
       (await asyncOptions()).overrides?.maxPriorityFeePerGas,
     nonce: async () => (await asyncOptions()).overrides?.nonce,
+    extraGas: async () => (await asyncOptions()).overrides?.extraGas,
   });
 }