+++ packages/thirdweb/src/extensions/erc20/__generated__/IVotes/write/delegateBySig.ts
@@ -191,5 +191,6 @@ export function delegateBySig(
     maxPriorityFeePerGas: async () =>
       (await asyncOptions()).overrides?.maxPriorityFeePerGas,
     nonce: async () => (await asyncOptions()).overrides?.nonce,
+    extraGas: async () => (await asyncOptions()).overrides?.extraGas,
   });
 }