+++ packages/thirdweb/src/extensions/erc4337/__generated__/IEntryPoint/write/handleOps.ts
@@ -205,5 +205,6 @@ export function handleOps(
     maxPriorityFeePerGas: async () =>
       (await asyncOptions()).overrides?.maxPriorityFeePerGas,
     nonce: async () => (await asyncOptions()).overrides?.nonce,
+    extraGas: async () => (await asyncOptions()).overrides?.extraGas,
   });
 }