+++ packages/thirdweb/src/extensions/erc721/__generated__/IERC721A/write/approve.ts
@@ -138,5 +138,6 @@ export function approve(
     maxPriorityFeePerGas: async () =>
       (await asyncOptions()).overrides?.maxPriorityFeePerGas,
     nonce: async () => (await asyncOptions()).overrides?.nonce,
+    extraGas: async () => (await asyncOptions()).overrides?.extraGas,
   });
 }