+++ packages/thirdweb/src/extensions/thirdweb/__generated__/IAppURI/write/setAppURI.ts
@@ -132,5 +132,6 @@ export function setAppURI(
     maxPriorityFeePerGas: async () =>
       (await asyncOptions()).overrides?.maxPriorityFeePerGas,
     nonce: async () => (await asyncOptions()).overrides?.nonce,
+    extraGas: async () => (await asyncOptions()).overrides?.extraGas,
   });
 }